Code review comment including move payment view creation logic
