+++ uisdk/src/main/java/com/karhoo/uisdk/screen/booking/booking/bookingrequest/BookingRequestMVP.kt
@@ -60,7 +60,7 @@ interface BookingRequestMVP {
         fun showPrebookConfirmationDialog(quoteType: QuoteType?, tripInfo: TripInfo)
 
         fun showUpdatedPaymentDetails(savedPaymentInfo: SavedPaymentInfo?)
-        
+
         fun hideLoading()
 
     }
@@ -87,8 +87,6 @@ interface BookingRequestMVP {
         fun showBookingRequest(quote: Quote, outboundTripId: String? = null)
 
         fun resetBooking()
-
-//        fun updateCardDetails(sdkNonce: String?)
 
         fun watchBookingStatus(bookingStatusStateViewModel: BookingStatusStateViewModel): Observer<in BookingStatus>
