+++ uisdk/src/main/java/com/karhoo/uisdk/screen/booking/booking/payment/BookingPaymentPresenter.kt
@@ -3,10 +3,13 @@ package com.karhoo.uisdk.screen.booking.booking.payment
 import com.karhoo.sdk.api.KarhooApi
 import com.karhoo.sdk.api.datastore.user.SavedPaymentInfo
 import com.karhoo.sdk.api.datastore.user.UserStore
+import com.karhoo.sdk.api.model.Provider
 import com.karhoo.sdk.api.network.response.Resource
 import com.karhoo.sdk.api.service.payments.PaymentsService
 import com.karhoo.uisdk.R
 import com.karhoo.uisdk.base.BasePresenter
+import com.karhoo.uisdk.screen.booking.booking.payment.adyen.AdyenPaymentView
+import com.karhoo.uisdk.screen.booking.booking.payment.braintree.BraintreePaymentView
 
 class BookingPaymentPresenter(view: BookingPaymentMVP.View,
                               private val userStore: UserStore = KarhooApi.userStore,
@@ -31,4 +34,26 @@ class BookingPaymentPresenter(view: BookingPaymentMVP.View,
             userStore.savedPaymentInfo = it
         }
     }
+
+    override fun createPaymentView(provider: Provider?, actions: PaymentDropInMVP.Actions) {
+        val paymentView = provider?.let {
+            when (enumValueOf<ProviderType>(it.id.toUpperCase())) {
+                ProviderType.ADYEN -> {
+                    val view = AdyenPaymentView(actions)
+                    view.actions = actions
+                    view
+                }
+                ProviderType.BRAINTREE -> {
+                    val view = BraintreePaymentView(actions)
+                    view.actions = actions
+                    view
+                }
+            }
+        }
+        view?.setPaymentView(view = paymentView)
+    }
+}
+
+enum class ProviderType {
+    ADYEN, BRAINTREE
 }
